+++ src/index.ts
@@ -115,6 +115,9 @@ export type { SliderProps } from "components/Slider";
 export { default as Switch } from "components/Switch";
 export type { SwitchProps } from "components/Switch";
 
+export { default as Table } from "components/Table";
+export type { TableProps } from "components/Table";
+
 export { default as Tabs } from "components/Tabs";
 export type { TabsProps } from "components/Tabs";