+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 2.4
 
 - SSR system/ls color mode
+- New component: Table
 
 - Tabs: auto scroll for active item